% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsExported.R
\name{changeirt}
\alias{changeirt}
\title{Compute RCI}
\usage{
changeirt(theta.t1, theta.t2, method = "wald")
}
\arguments{
\item{method}{character: the test used; defaults to Wald test}

\item{theta_pre}{matrix: first column thetas, second column se}

\item{theta_post}{matrix: first column thetas, second column se}
}
\value{
children of class ChangeAssessment (to be added...)
}
\description{
Roxygen2 is awesome.
}
