% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{estimateirt}
\alias{estimateirt}
\title{Estimate IRT model based on data}
\usage{
estimateirt(data)
}
\arguments{
\item{data}{matrix}
}
\value{
children of class Estimator (to be added...)
}
\description{
Works for 2PLM and GRM - decide the dependency: ltm or mirt?
}
