% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsExported.R
\name{personfit}
\alias{personfit}
\title{Compute Person-Fit statistics}
\usage{
personfit()
}
\value{
matrix: first column test value, second column p value
}
\description{
to be added... decide: dependecy or implementation
}
