% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericDefinitions.R, R/GenericMethods.R
\docType{methods}
\name{simulateirt}
\alias{simulateirt}
\alias{simulateirt,numeric,numeric,missing,missing,missing-method}
\alias{simulateirt}
\alias{simulateirt,numeric,numeric,numeric,missing,missing-method}
\alias{simulateirt}
\alias{simulateirt,numeric,numeric,numeric,numeric,missing-method}
\alias{simulateirt}
\alias{simulateirt,numeric,numeric,numeric,numeric,numeric-method}
\alias{simulateirt}
\title{Method simulateirt}
\usage{
simulateirt(theta, difficulty, discrimination, lower_asymptote, upper_asymptote)

\S4method{simulateirt}{numeric,numeric,missing,missing,missing}(theta,
  difficulty)

\S4method{simulateirt}{numeric,numeric,numeric,missing,missing}(theta,
  difficulty, discrimination)

\S4method{simulateirt}{numeric,numeric,numeric,numeric,missing}(theta,
  difficulty, discrimination, lower_asymptote)

\S4method{simulateirt}{numeric,numeric,numeric,numeric,numeric}(theta,
  difficulty, discrimination, lower_asymptote, upper_asymptote)
}
\arguments{
\item{theta}{A numerical vector containing the theta values to be used in the simulation. The \code{length(theta)} gives the number of respondents (e.g., the number of rows in the matrix of item responses).}

\item{difficulty}{A numerical vector containing the difficulty item parameters (also known as "b"). If \code{length(difficulty)} is grater than \code{0} and identical with the length of all the other item paramters, it gives the numer of items (e.g., the number of rows in the matrix of item responses).}

\item{discrimination}{A numerical vector containing the discriminability item parameters (also known as "a").}

\item{lower_asymptote}{A numerical vector containing the lower asymptote item parameters (also known as "c").}

\item{upper_asymptote}{A numerical vector containing the upper asymptote item parameters (also known as "d").}
}
\value{
An object of class \code{\link{Simulator}}. Note that the object implements the formal S4 class system \url{http://adv-r.had.co.nz/S4.html}. Therefore, the generated data is contained within different fields named \code{slot}. To access, for instance, details about the items use \code{SimulatorObject@item}. This will print all the elements within the list \code{item}. Subsequent access to thee elements is permited using the \code{$} operator, e.g, \code{SimulatorObject@item$parameters$difficulty}. See \code{\link{Simulator}} class structure for an idea of what can be accessed.
}
\description{
Method \code{simulateirt} is the end-user interface for generating dichotomous item responses. It is a generic function. 
It currently supports "1PL", "2PL", "3PL", and "4PL" IRT models. It automatically determines the appropriate IRT model based on the arguments passed. 
There are 4 models supported, therefore there are 4 use cases. It is highly recommended to use named arguments to make sure the proper model is used. Use cases:
\itemize{
\item 1PL model is used if signature matches: \code{simulateirt(theta = vector, difficulty = vector)}
\item 2PL model is used if signature matches: \code{simulateirt(theta = vector, difficulty = vector, discriminability = vector)}
\item 3PL model is used if signature matches: \code{simulateirt(theta = vector, difficulty = vector, discriminability = vector, lower_asymptote = vector)}
\item 4PL model is used if signature matches: \code{simulateirt(theta = vector, difficulty = vector, discriminability = vector, lower_asymptote = vector, upper_asymptote = vector)}
}
}
