% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsExported.R
\name{simulateirt}
\alias{simulateirt}
\title{Simulate IRT data}
\usage{
simulateirt(theta, b, a)
}
\arguments{
\item{theta}{vector}

\item{b}{vector or matrix (if matrix GRM is used)}

\item{a}{vector (of equal size with \code{length(b)})}
}
\value{
children of class Simulator (to be added...)
}
\description{
Works for 2PLM and GRM
}
