% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsExported.R
\name{changeirt}
\alias{changeirt}
\title{Compute RCI}
\usage{
changeirt(theta.t1, theta.t2, method = "wald")
}
\arguments{
\item{theta.t1}{matrix: first column thetas, second column se}

\item{theta.t2}{matrix: first column thetas, second column se}

\item{method}{character: the test used; defaults to Wald test}
}
\value{
children of class ChangeAssessment (to be added...)
}
\description{
to be added...
}
